修改增加资金明细俩rpc

+++ src/main/java/com/dx/facade/texas/req/TableBringInfoReq.java
@@ -12,7 +12,7 @@ import java.util.Set;
 @Data
 public class TableBringInfoReq implements Serializable, BaseParmDTO {
   @ApiModelProperty("牌桌ids")
-  private Set<Long> tableIds;
+  private Set<Integer> tableIds;
 
   @Override
   public void check() {
